+++ internal/app/user_favorite_brand/user_favorite_brand_cache.go
@@ -28,7 +28,7 @@ func NewUserFavoriteBrandCacheRepository(repo interfaces.UserFavoriteBrandReposi
 	}
 }
 
-func (c userFavoriteBrandCache) GetListFavBrandByUserIdAndBrandIds(ctx context.Context, userId uint64, brandIds []uint) ([]model.UserFavoriteBrand, error) {
+func (c userFavoriteBrandCache) GetListFavBrandByUserIdAndBrandIds(ctx context.Context, userId uint64, brandIds []uint64) ([]model.UserFavoriteBrand, error) {
 	s, _ := json.Marshal(brandIds)
 	key := &caching.Keyer{Raw: keyPrefixUserFavoriteBrand + fmt.Sprint("GetListFavBrandByUserIdAndBrandIds_", userId, "_", string(s))}
 	var listUserFavoriteBrand []model.UserFavoriteBrand
